moved environment.ts inside environment folder
