+++ src/app/services/settings-broadcasting.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { environment } from 'src/environments.ts/environment';
+import { environment } from 'src/environments/environment';
 
 @Injectable({
   providedIn: 'root',
